\name{Mondrian-package}

\alias{Mondrian-package}
\alias{Mondrian}

\docType{package}

\title{Display the relative occurrence of multi-infectious statuses}

\description{
This package comes down to one function which represents the relative occurrence of individual multi-infectious statuses.
The main graphical output looks like Mondrian paintings.
}

\details{
  \tabular{ll}{
    Package: \tab Mondrian\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2015-05-05\cr
    License: \tab GPL (>=2)\cr
  }
}

\author{
  Aurelie Siberchicot <aurelie.siberchicot@univ-lyon1.fr>, Delphine Charif, Gabriel Terraz and Fabrice Vavre
}

\keyword{package}

\seealso{
  \code{\link{mondrian}}
}
