\name{Mondrian-package}

\alias{Mondrian-package}
\alias{Mondrian}

\docType{package}

\title{A Simple Graphical Representation of the Relative Occurrence and Co-Occurrence of Events}

\description{The unique function of this package allows representing in a single graph the relative occurrence and co-occurrence of events measured in a sample.\cr
  As examples, the package was applied to describe the occurrence and co-occurrence of different species of bacterial or viral symbionts infecting 
  arthropods at the individual level. The graphics allows determining the prevalence of each symbiont and the patterns of multiple infections 
  (i.e. how different symbionts share or not the same individual hosts). \cr
  We named the package after the famous painter as the graphical output recalls Mondrian’s famous paintings.
}

\details{
  \tabular{ll}{
    Package: \tab Mondrian\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2016-01-07\cr
    License: \tab GPL (>=2)\cr
  }
}

\author{
  Aurélie Siberchicot <aurelie.siberchicot@univ-lyon1.fr>, Delphine Charif, Gabriel Terraz and Fabrice Vavre
}

\keyword{package}

\seealso{
  \code{\link{mondrian}}
}
